[CWS] Code cleanup
