--- pkg/clusteragent/admission/mutate/cwsinstrumentation/k8scp/exec.go
@@ -1,198 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build kubeapiserver
-
-package k8scp
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"net/url"
-	"time"
-
-	dockerterm "github.com/moby/term"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/cli-runtime/pkg/genericiooptions"
-	coreclient "k8s.io/client-go/kubernetes/typed/core/v1"
-	restclient "k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/remotecommand"
-
-	"k8s.io/kubectl/pkg/cmd/util/podcmd"
-	"k8s.io/kubectl/pkg/scheme"
-	"k8s.io/kubectl/pkg/util/interrupt"
-	"k8s.io/kubectl/pkg/util/term"
-)
-
-// RemoteExecutor defines the interface accepted by the Exec command - provided for test stubbing
-type RemoteExecutor interface {
-	Execute(url *url.URL, config *restclient.Config, stdin io.Reader, stdout, stderr io.Writer, tty bool, terminalSizeQueue remotecommand.TerminalSizeQueue) error
-}
-
-// DefaultRemoteExecutor is the standard implementation of remote command execution
-type DefaultRemoteExecutor struct{}
-
-// Execute runs with the provided config
-func (*DefaultRemoteExecutor) Execute(url *url.URL, config *restclient.Config, stdin io.Reader, stdout, stderr io.Writer, tty bool, terminalSizeQueue remotecommand.TerminalSizeQueue) error {
-	// Legacy SPDY executor is default. If feature gate enabled, fallback
-	// executor attempts websockets first--then SPDY.
-	exec, err := remotecommand.NewSPDYExecutor(config, "POST", url)
-	if err != nil {
-		return err
-	}
-	return exec.StreamWithContext(context.Background(), remotecommand.StreamOptions{
-		Stdin:             stdin,
-		Stdout:            stdout,
-		Stderr:            stderr,
-		Tty:               tty,
-		TerminalSizeQueue: terminalSizeQueue,
-	})
-}
-
-// StreamOptions contains option for the remote stream
-type StreamOptions struct {
-	Namespace     string
-	PodName       string
-	ContainerName string
-	Stdin         bool
-	TTY           bool
-	// minimize unnecessary output
-	Quiet bool
-	// InterruptParent, if set, is used to handle interrupts while attached
-	InterruptParent *interrupt.Handler
-
-	genericiooptions.IOStreams
-
-	// for testing
-	overrideStreams func() (io.ReadCloser, io.Writer, io.Writer)
-	isTerminalIn    func(t term.TTY) bool
-}
-
-// ExecOptions declare the arguments accepted by the Exec command
-type ExecOptions struct {
-	StreamOptions
-
-	ResourceName     string
-	Command          []string
-	EnforceNamespace bool
-
-	Pod           *corev1.Pod
-	Executor      RemoteExecutor
-	PodClient     coreclient.PodsGetter
-	GetPodTimeout time.Duration
-	Config        *restclient.Config
-}
-
-// SetupTTY sets up the TTY
-func (o *StreamOptions) SetupTTY() term.TTY {
-	t := term.TTY{
-		Parent: o.InterruptParent,
-		Out:    o.Out,
-	}
-
-	if !o.Stdin {
-		// need to nil out o.In to make sure we don't create a stream for stdin
-		o.In = nil
-		o.TTY = false
-		return t
-	}
-
-	t.In = o.In
-	if !o.TTY {
-		return t
-	}
-
-	if o.isTerminalIn == nil {
-		o.isTerminalIn = func(tty term.TTY) bool {
-			return tty.IsTerminalIn()
-		}
-	}
-	if !o.isTerminalIn(t) {
-		o.TTY = false
-
-		if !o.Quiet && o.ErrOut != nil {
-			fmt.Fprintln(o.ErrOut, "Unable to use a TTY - input is not a terminal or the right kind of file")
-		}
-
-		return t
-	}
-
-	// if we get to here, the user wants to attach stdin, wants a TTY, and o.In is a terminal, so we
-	// can safely set t.Raw to true
-	t.Raw = true
-
-	if o.overrideStreams == nil {
-		// use dockerterm.StdStreams() to get the right I/O handles on Windows
-		o.overrideStreams = dockerterm.StdStreams
-	}
-	stdin, stdout, _ := o.overrideStreams()
-	o.In = stdin
-	t.In = stdin
-	if o.Out != nil {
-		o.Out = stdout
-		t.Out = stdout
-	}
-
-	return t
-}
-
-// Run executes a validated remote execution against a pod.
-func (p *ExecOptions) Run(pod *corev1.Pod) error {
-	if pod.Status.Phase == corev1.PodSucceeded || pod.Status.Phase == corev1.PodFailed {
-		return fmt.Errorf("cannot exec into a container in a completed pod; current phase is %s", pod.Status.Phase)
-	}
-
-	containerName := p.ContainerName
-	if len(containerName) == 0 {
-		container, err := podcmd.FindOrDefaultContainerByName(pod, containerName, p.Quiet, p.ErrOut)
-		if err != nil {
-			return err
-		}
-		containerName = container.Name
-	}
-
-	// ensure we can recover the terminal while attached
-	t := p.SetupTTY()
-
-	var sizeQueue remotecommand.TerminalSizeQueue
-	if t.Raw {
-		// this call spawns a goroutine to monitor/update the terminal size
-		sizeQueue = t.MonitorSize(t.GetSize())
-
-		// unset p.Err if it was previously set because both stdout and stderr go over p.Out when tty is
-		// true
-		p.ErrOut = nil
-	}
-
-	fn := func() error {
-		restClient, err := restclient.RESTClientFor(p.Config)
-		if err != nil {
-			return err
-		}
-
-		req := restClient.Post().
-			Resource("pods").
-			Name(pod.Name).
-			Namespace(pod.Namespace).
-			SubResource("exec")
-		req.VersionedParams(&corev1.PodExecOptions{
-			Container: containerName,
-			Command:   p.Command,
-			Stdin:     p.Stdin,
-			Stdout:    p.Out != nil,
-			Stderr:    p.ErrOut != nil,
-			TTY:       t.Raw,
-		}, scheme.ParameterCodec)
-
-		return p.Executor.Execute(req.URL(), p.Config, p.In, p.Out, p.ErrOut, t.Raw, sizeQueue)
-	}
-
-	if err := t.Safe(fn); err != nil {
-		return err
-	}
-
-	return nil
-}
