+++ pkg/clusteragent/admission/mutate/cwsinstrumentation/k8scp/streamOptions.go
@@ -1,0 +1,67 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build kubeapiserver
+
+package k8scp
+
+import (
+	"fmt"
+
+	dockerterm "github.com/moby/term"
+	"k8s.io/cli-runtime/pkg/genericiooptions"
+	"k8s.io/kubectl/pkg/util/term"
+)
+
+// StreamOptions contains option for the remote stream
+type StreamOptions struct {
+	Stdin bool
+	TTY   bool
+
+	genericiooptions.IOStreams
+}
+
+// SetupTTY sets up the TTY
+func (o *StreamOptions) SetupTTY() term.TTY {
+	t := term.TTY{
+		Out: o.Out,
+	}
+
+	if !o.Stdin {
+		// need to nil out o.In to make sure we don't create a stream for stdin
+		o.In = nil
+		o.TTY = false
+		return t
+	}
+
+	t.In = o.In
+	if !o.TTY {
+		return t
+	}
+
+	if !t.IsTerminalIn() {
+		o.TTY = false
+
+		if o.ErrOut != nil {
+			fmt.Fprintln(o.ErrOut, "Unable to use a TTY - input is not a terminal or the right kind of file")
+		}
+
+		return t
+	}
+
+	// if we get to here, the user wants to attach stdin, wants a TTY, and o.In is a terminal, so we
+	// can safely set t.Raw to true
+	t.Raw = true
+
+	stdin, stdout, _ := dockerterm.StdStreams()
+	o.In = stdin
+	t.In = stdin
+	if o.Out != nil {
+		o.Out = stdout
+		t.Out = stdout
+	}
+
+	return t
+}
